Add files via upload
In this project, I used a multiple regression mechanism to identify the best-fitting machine learning model to the data set by Hyper-parameter tuning using Grid search cross-validation. Then later, the best-fitting ML model is trained for forecasting. The model is then saved as a pkl file for later deployment.
Therefore, the project is divided into the following processes.
- Data Preparation
- Data Preprocessing
- Data Visualization
- Modelling
- Evaluation and Choosing the Best-Fit model
- Forecasting Predictions
- Saving, Deleting and Loading the BestFitModel